requirements.txt
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
curl_cffi==0.13.0
exceptiongroup==1.3.0
Flask==3.1.2
frozendict==2.4.6
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multitasking==0.0.12
numpy==2.2.6
packaging==25.0
pandas==2.3.3
platformdirs==4.4.0
pluggy==1.6.0
protobuf==6.32.1
pycparser==2.23
Pygments==2.19.2
pytest==8.4.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
soupsieve==2.8
tomli==2.2.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
yfinance==0.2.66
cloud-sql-python-connector
pg8000
yoyo-migrations